+++ src/onegov.gazette/onegov/gazette/tests/conftest.py
@@ -81,23 +81,27 @@ def create_gazette(request):
         role='admin'
     ))
     app.session().add(User(
+        realname='Publisher',
         username='publisher@example.org',
         password_hash=request.getfixturevalue('gazette_password'),
         role='editor'
     ))
     app.session().add(User(
+        realname='First Editor',
         username='editor1@example.org',
         password_hash=request.getfixturevalue('gazette_password'),
         role='member',
         group_id=group_id
     ))
     app.session().add(User(
+        realname='Second Editor',
         username='editor2@example.org',
         password_hash=request.getfixturevalue('gazette_password'),
         role='member',
         group_id=group_id
     ))
     app.session().add(User(
+        realname='Third Editor',
         username='editor3@example.org',
         password_hash=request.getfixturevalue('gazette_password'),
         role='member'